# parking/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: Parking api
host: 'localhost:10010'
basePath: /v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /admin/update:
    post:
      summary: Activa o desabilita un usuario
      description: Permite cambiar el estado de un usuario 
      tags:
        - User
      x-swagger-router-controller: user
      operationId: updatesUser
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: Informacion sobre el usuario que quiere ser activado o desabilitado, solo es necesario enviar el rut del usuario y el estado nuevo 1 para ser habilitado, 0 para ser desactivado
          schema:
            $ref: '#/definitions/userInfo'
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
...
  /user/estacionamiento:
    get:
      summary: obtiene los estacionamientos del usuario
      description: Permite la obtener todos los estacionamientos  del usuario
      tags:
        - User
        - Estacionamiento
      x-swagger-router-controller: user
      operationId: findEstacionamientos
      parameters:
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/estacionamientosUser'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /user/auto:
    get:
      summary: obtiene los auto del usuario
      description: Permite la obtener todos los auto  del usuario
      tags:
        - User
        - Auto
      x-swagger-router-controller: user
      operationId: findAutos
      parameters:
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/autosUser'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
  /user/reservas:
    get:
      summary: obtiene los reservas del usuario
      description: Permite la obtener todos las reservas del usuario
      tags:
        - User
        - Reserva
      x-swagger-router-controller: user
      operationId: findReservas
      parameters:
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/reservasUser'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []        
  /user/pagos:
    get:
      summary: obtiene los pagos del usuario
      description: Permite la obtener todos las pagos del usuario
      tags:
        - User
        - Reserva
      x-swagger-router-controller: user
      operationId: findPagos
      parameters:
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/pagosUser'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []   
  /user/pendientes:
    get:
      summary: obtiene las calificaciones pendientes del usuario
      description: Permite las calificaciones pendientes  del usuario
      tags:
        - User
        - Calificación
      x-swagger-router-controller: user
      operationId: findPendientes
      parameters:
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/pendientesUser'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: [] 
  /user/rate:
    post:
      summary: Permite calificar un usuario
      description: Permite la calificacion de un usuario en el sistema
      tags:
        - User
        - Calificación
      x-swagger-router-controller: user
      operationId: createRate
      parameters:
        - in: body
          name: body
          description: ''
          required: true
          schema:
            $ref: '#/definitions/rate'
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: [] 
  /user/update:
    get:
      summary: cambia el rol del  usuario a dueño
      description: Permite cambiar el rol del usuario de cliente a dueño
      tags:
        - User
      x-swagger-router-controller: user
      operationId: updateRol
      parameters:
        - in: header
          name: api_key
          description: ''
          required: true
          type: string
      responses:
        '200':
          description: Operación exitosa
          schema:
            $ref: '#/definitions/mensajeModel'
        '400':
          description: Solicitud incorrecta
          schema:
            $ref: '#/definitions/mensajeModel'
        '403':
          description: Credenciales incorrectas
          schema:
            $ref: '#/definitions/mensajeModel'
        '500':
          description: Error en sistema
          schema:
            $ref: '#/definitions/mensajeModel'
      security:
        - api_key: []
...
definitions:
  mensajeModel:
    type: object
    properties:
      message:
        type: string
      code:
        type: string
...
  userInfo:
    type: object
    properties:
      rut:
        type: integer
        format: int64
      div:
        type: string
      nombre:
        type: string
      apPaterno:
        type: string
      apMaterno:
        type: string
      correo:
        type: string
      telefono:
        type: integer
        format: int64
      estado:
        type: integer
        format: int64     
...
  reservasUser:
    type: array
    items:
      $ref: '#/definitions/reserva'
  auto:
    type: object
    properties:
      patente:
        type: string
      marca:
        type: string
      color:
        type: string
      motor:
        type: string
      chasis:
        type: string
...
  estacionamiento:
    type: object
    properties:
      id:
        description: Id del estacionamiento, la cual no se envia al momento de crear un estacionamiento
        type: integer
        format: int64
      horarios:
        description: indica la cantidad de horararios que tiene activos el estacionamiento
        type: integer
        format: int64
      costo:
        type: integer
        format: int64
      geo:
        $ref: '#/definitions/geo'
      comunaId:
        type: integer
        format: int64
      estado:
        type: integer
        format: int64
  rate:
    type: object
    properties:
      id:
        description: Id de la calificación
        type: integer
        format: int64
      comentario:
        type: string   
      rate:
        description: Cantidad de estrellas del 1 al 5
        type: integer
        format: int64
...
  autosUser:
    type: array
    items:
      $ref: '#/definitions/auto' 
  pagosUser:
    type: array
    items:
      $ref: '#/definitions/pago' 
...
  pendientesUser:
    type: array
    items:
      $ref: '#/definitions/pendiente' 
  estacionamientosUser:
    type: array
    items:
      $ref: '#/definitions/estacionamiento'
...
  pendiente:
    type: object
    properties:
      id:
        type: integer
        format: int64
      horaEntrada:
        type: string
        format: dateTime
      horaSalida:
        type: string
        format: dateTime
      idReserva:
        type: integer
        format: int64
      nombre:
        type: string
      apPaterno:
        type: string
      apMaterno:
        type: string
  pago:
    type: object
    properties:
      horaEntrada:
        type: string
        format: dateTime
      horaSalida:
        type: string
        format: dateTime
      estadoPago:
        description: puede ser del tipo Aprovado  o rechazado
        type: string
      total:
        type: integer
        format: int64
      type:
        description: indica que tipo de pago se cancelo
        type: string
...
  reserva:
    type: object
    properties:
      id:
        type: integer
        format: int64
      horaEntrada:
        type: string
        format: dateTime
      horaSalida:
        type: string
        format: dateTime
      fecha:
        type: string
        format: dateTime
      total:
        type: integer
        format: int64
      autoPatente:
        type: string
      telefono:
        type: integer
        format: int64
      correo:
        type: string
      nombre:
        type: string
      apPaterno:
        type: string
      apMaterno:
        type: string
      idHorario:
        type: integer
        format: int64
  geo:
    type: object
    properties:
      lat:
        type: number
        format: float
      lon:
        type: number
        format: float
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header